Add test of large-scale query syntax
This is preparatory to updating to ES 6.x. We'll change this test, in
consultation with the changelog, to the new format, and then we can
update model behavior until it passes.

While we're at it, convert some strings to symbols, and use them
consistently throughout. Yes, Hashie lets us get away with the
differences in most cases (not in the equality assertion in the test,
though!)